Fix Work Order Quote table/UDT mismatches
- WorkOrderQuoteChargesType.Quantity: INT → DECIMAL(18,6) to match table
- WorkOrderQuoteDetailsType: 31 financial columns DECIMAL(20,2) → DECIMAL(18,6)
  and CommonFlatRate DECIMAL(9,2) → DECIMAL(18,6) to match table precision
- WorkOrderQuoteLaborType: 6 financial columns DECIMAL(20,2) → DECIMAL(18,6)
  (DirectLaborOHCost, BurdenRateAmount, TotalCostPerHour, TotalCost,
   BillingRate, BillingAmount) to match table precision
- WorkOrderQuoteMaterialType.UomName: VARCHAR(50) → VARCHAR(100) to match table
- WorkOrderQuoteTask: add IsKitPart BIT NULL (present in UDT, missing in table)
- WorkOrderQuoteTaskAudit: add IsKitPart BIT NULL to keep trigger SELECT * in sync

+++ PAS_Database/dbo/User Defined Types/WorkOrderQuoteChargesType.sql
@@ -3,7 +3,7 @@
     [WorkOrderQuoteDetailsId] BIGINT          NULL,
     [ChargesTypeId]           BIGINT          NULL,
     [VendorId]                BIGINT          NULL,
-    [Quantity]                INT             NULL,
+    [Quantity]                DECIMAL (18, 6) NULL,
     [MarkupPercentageId]      BIGINT          NULL,
     [Description]             VARCHAR (256)   NULL,
     [UnitCost]                DECIMAL (20, 2) NULL,
